refactor: move probe fixtures to own file
