+++ src/components/ProbeList/ProbeList.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
-import { DEFAULT_PROBES } from 'test/fixtures';
+import { DEFAULT_PROBES } from 'test/fixtures/probes';
 import { render } from 'test/render';
 
 import { ProbeList } from './ProbeList';
