+++ src/page/Probes/Probes.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { screen, within } from '@testing-library/react';
 import { DataTestIds } from 'test/dataTestIds';
-import { DEFAULT_PROBES, OFFLINE_PROBE, ONLINE_PROBE, PRIVATE_PROBE, PUBLIC_PROBE } from 'test/fixtures';
+import { DEFAULT_PROBES, OFFLINE_PROBE, ONLINE_PROBE, PRIVATE_PROBE, PUBLIC_PROBE } from 'test/fixtures/probes';
 import { render } from 'test/render';
 
 import { ROUTES } from 'types';
